// frontend/src/gen/types/CreateRouteRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type CreateRouteRequest = {
  /**
   * @description List of destination points for the route
   * @type array
   */
  points: string[];
  /**
   * @description Start date for the trip in YYYY-MM-DD format
   * @type string, date
   */
  startDate: string;
  /**
   * @description Minimum duration of the trip in days
   * @minLength 1
   * @type integer
   */
  durationMinDays: number;
  /**
   * @description Maximum duration of the trip in days (optional)
   * @minLength 1
   * @type integer
   */
  durationMaxDays?: number | null;
};